requirements.txt
Django==2.2
django-environ==0.4.5
django-model-utils==4.0.0
pkg-resources==0.0.0
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1